Move get_mean_mrh() to geometry as it fits better there.
